added validations

+++ Calculator/Operations/Operation.cs
@@ -4,16 +4,15 @@
     {
         public double X { get; set; }
         public double Y { get; set; }
-
         public Operation(double x, double y)
         {
             X = x;
             Y = y;
         }
 
-        public Operation(double x)
+        public Operation(int x)
         {
-            X = x;            
+            X = x;
         }
 
         public abstract double toResult();
